Interruptible poll()
When poll() is waiting for events on watched file descriptors, it should
not block any thread wanting to (un)schedule an alarm.

But the mutex cannot be released while executing poll(), because the
watched file descriptors must not be modified (or closed!) by another
thread.

Therefore, poll() now waits for one more file descriptor, a pipe, to
become ready.

Before acquiring a mutex, fdqueue functions write 1 byte to the pipe.
Once the mutex is acquired, they read 1 byte. Thus, poll() is guaranteed
to be non-blocking when a thread waits for the mutex.

If the pipe is ready for I/O, then fd_poll() release the mutex for 1ms.