Multiple fdqueues support
The purpose is to execute rhizome actions on a thread different from
that handling routing and MDP stuff, so that a long rhizome action does
not jam the communications.

This first step adds a "struct fdqueue" type with all the fields
specific to an fdqueue. Two instances are created: main_fdqueue and
rhizome_fdqueue.

All public fdqueue functions now take an fdqueue argument. They are
thread-safe.

fd_poll() uses synchronization privimives for waiting until the next
event, so that the wait can be interrupted for executing a new action
earlier.