Modify OpenWRT Makefile for development
Now that OpenWRT Makefiles for the OpenWRT release feeds are in the
servalproject/openwrt-packages repository, there should be no OpenWRT release
Makefiles anywhere else.  So the Makefile that was in the Serval DNA repo now
serves for developers who want to edit Serval DNA source and compile for
OpenWRT without having to commit and push to GitHub before every compile.