Fix a non-deterministic server test case failure
Very occasionally, when the file logger was creating the symlink from
'serval.log' to the current log file, there was a race with another
process doing exactly the same, causing one of them to log an
symlink-failed ERROR that caused the server/StartTwice test to fail
(when asserting that the "start" command contained no ERROR log
messages).  Now the symlink failure is logged at WARN level, which will
not cause the test to fail.